Merge recent ExoPlayer releases

+++ library/extractor/src/main/java/com/google/android/exoplayer2/extractor/mp3/VbriSeeker.java
@@ -55,8 +55,9 @@ import com.google.android.exoplayer2.util.Util;
       return null;
     }
     int sampleRate = mpegAudioHeader.sampleRate;
-    long durationUs = Util.scaleLargeTimestamp(numFrames,
-        C.MICROS_PER_SECOND * (sampleRate >= 32000 ? 1152 : 576), sampleRate);
+    long durationUs =
+        Util.scaleLargeTimestamp(
+            numFrames, C.MICROS_PER_SECOND * (sampleRate >= 32000 ? 1152 : 576), sampleRate);
     int entryCount = frame.readUnsignedShort();
     int scale = frame.readUnsignedShort();
     int entrySize = frame.readUnsignedShort();
@@ -88,7 +89,7 @@ import com.google.android.exoplayer2.util.Util;
         default:
           return null;
       }
-      position += segmentSize * scale;
+      position += segmentSize * ((long) scale);
     }
     if (inputLength != C.LENGTH_UNSET && inputLength != position) {
       Log.w(TAG, "VBRI data size mismatch: " + inputLength + ", " + position);
